Fixed a compiler warning in compound_id_v0.cpp.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@59712 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/compound_id_v0.cpp
@@ -360,7 +360,7 @@ CCompoundID SIDUnpacking::ExtractCID(SCompoundIDPoolImpl* pool_impl)
             new_cid.AppendInteger(ExtractNumber());
             break;
         case '-':
-            new_cid.AppendInteger(-ExtractNumber());
+            new_cid.AppendInteger(-(Int8) ExtractNumber());
             break;
         case CIT_SERVICE_NAME_FIELD_CODE:
             new_cid.AppendServiceName(ExtractString());
